refactor: 인가 -> 인증 용어 변경
기존에는 로그인이 필요한 서비스에 인가라는 표현을 썼지만,
인증이 맞기 때문에 용어를 Authoriaztion -> Authentication으로 변경

현재 인증 로직은 2가지 입니다.
1. 소셜 로그인 인증
2. 글 작성과 같은 로그인이 필요한 인증
