+++ src/main/java/project/myblog/auth/authentication/Authentication.java
@@ -3,7 +3,7 @@ package project.myblog.auth.authentication;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public interface Logout {
+public interface Authentication {
     boolean isSupported(HttpServletRequest request, HttpServletResponse response);
-    void logout(HttpServletRequest request, HttpServletResponse response);
+    void authorize(HttpServletRequest request, HttpServletResponse response);
 }
