+++ src/main/java/project/myblog/auth/oauthAuthentication/Logout.java
@@ -1,9 +1,9 @@
-package project.myblog.auth.authorization;
+package project.myblog.auth.oauthAuthentication;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public interface Authorization {
+public interface Logout {
     boolean isSupported(HttpServletRequest request, HttpServletResponse response);
-    void authorize(HttpServletRequest request, HttpServletResponse response);
+    void logout(HttpServletRequest request, HttpServletResponse response);
 }
