+++ src/main/java/project/myblog/exception/ExceptionCode.java
@@ -9,7 +9,7 @@ import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 
 public enum ExceptionCode {
     MEMBER_INVALID(BAD_REQUEST, "MEMBER_001", "존재하지 않는 회원입니다."),
-    MEMBER_AUTHORIZATION(FORBIDDEN, "MEMBER_002", "로그인이 필요합니다."),
+    MEMBER_AUTHORIZATION(UNAUTHORIZED, "MEMBER_002", "로그인 인증이 필요합니다."),
     MEMBER_AUTHENTICATION(UNAUTHORIZED, "MEMBER_003", "인증되지 않는 사용자입니다."),
     POST_INVALID(NOT_FOUND, "POST_001", "존재하지 않는 포스트입니다."),
     POST_AUTHORIZATION(FORBIDDEN, "POST_002", "포스트에 대한 권한이 없습니다.");
